+++ familyhelper-project-sdk/src/main/java/com/dwarfeng/familyhelper/project/sdk/util/ServiceExceptionCodes.java
@@ -16,14 +16,20 @@ public final class ServiceExceptionCodes {
             new ServiceException.Code(offset(0), "asset catalog not exists");
     public static final ServiceException.Code USER_NOT_EXISTS =
             new ServiceException.Code(offset(10), "user not exists");
-    public static final ServiceException.Code USER_NOT_PERMITTED =
-            new ServiceException.Code(offset(20), "user not permitted");
+    public static final ServiceException.Code USER_NOT_PERMITTED_FOR_PROJECT =
+            new ServiceException.Code(offset(20), "user not permitted for project");
     public static final ServiceException.Code INVALID_PERMISSION_LEVEL =
             new ServiceException.Code(offset(30), "invalid permission level");
     public static final ServiceException.Code INVALID_PROJECT_STATUS =
             new ServiceException.Code(offset(40), "invalid project status");
     public static final ServiceException.Code POST_TASK_EXISTS =
             new ServiceException.Code(offset(50), "post task exists");
+    public static final ServiceException.Code INVALID_MEMO_STATUS =
+            new ServiceException.Code(offset(60), "invalid memo status");
+    public static final ServiceException.Code MEMO_NOT_EXISTS =
+            new ServiceException.Code(offset(70), "memo not exists");
+    public static final ServiceException.Code USER_NOT_IDENTICAL =
+            new ServiceException.Code(offset(80), "post task exists");
 
     private static int offset(int i) {
         return EXCEPTION_CODE_OFFSET + i;
@@ -50,10 +56,13 @@ public final class ServiceExceptionCodes {
         // 以新的 EXCEPTION_CODE_OFFSET 为基准，更新异常代码的值。
         PROJECT_NOT_EXISTS.setCode(offset(0));
         USER_NOT_EXISTS.setCode(offset(10));
-        USER_NOT_PERMITTED.setCode(offset(20));
+        USER_NOT_PERMITTED_FOR_PROJECT.setCode(offset(20));
         INVALID_PERMISSION_LEVEL.setCode(offset(30));
         INVALID_PROJECT_STATUS.setCode(offset(40));
         POST_TASK_EXISTS.setCode(offset(50));
+        INVALID_MEMO_STATUS.setCode(60);
+        MEMO_NOT_EXISTS.setCode(70);
+        USER_NOT_IDENTICAL.setCode(80);
     }
 
     private ServiceExceptionCodes() {